Isolate Ensu storage compatibility code
