+++ rust/apps/ensu/changes/simplify-local-chat-storage.md
@@ -1,0 +1,1 @@
+- Simplified local chat storage while preserving existing chats and attachments.
