+++ web/apps/ensu/src/services/chat/compatibility.ts
@@ -1,15 +1,14 @@
 import { base64ToBytes } from "@/services/base64";
-import { secureStorageGet } from "@/services/secure-storage";
 import { isTauriRuntime } from "@/services/tauri-runtime";
+import { removeKV } from "ente-base/kv";
 import log from "ente-base/log";
 import { deleteDB, openDB, type DBSchema } from "idb";
-import { decryptAttachmentBytes, encryptAttachmentBytes } from "./attachments";
-import { localChatMigrationSourceKeys } from "./chatKey";
 import {
-    finalizeLocalChatMigration,
-    isLocalChatMigrationDone,
-    LEGACY_ATTACHMENT_SECURE_KEY,
-} from "./keyMigration";
+    secureStorageDelete,
+    secureStorageGet,
+    secureStorageSet,
+} from "../secure-storage";
+import { decryptAttachmentBytes, encryptAttachmentBytes } from "./attachments";
 import {
     attachmentBytesExists,
     chatDb,
@@ -28,6 +27,11 @@ import {
 const LOCAL_STORAGE_STORE = "ensu.chat.store.v1";
 const TOMBSTONE_MIGRATION = "ensu.chat.localMigration.v2";
 const OLD_INDEXED_DB = "ensu.chat.db";
+const MIGRATION_DONE = "ensu.chat.localMigration.v3";
+const OLD_LOCAL_SECURE_KEY = "localChatKey";
+const OLD_LOCAL_BROWSER_KEY = "ensu.chatKey.local";
+const KEY_FILE = "chat-keys.json";
+const LEGACY_ATTACHMENT_SECURE_KEY = "legacyAttachmentKey.v2";
 
 type OldSession = StoredSession & {
     deletedAt?: number | null;
@@ -51,7 +55,118 @@ interface OldIndexedDb extends DBSchema {
     attachmentBytes: { key: string; value: AttachmentBytes };
 }
 
-let _initPromise: Promise<void> | undefined;
+let _migrationSourceKeys: string[] = [];
+
+const localGet = (key: string) => localStorage.getItem(key) ?? undefined;
+
+const keyFilePath = async () => {
+    const { appDataDir, join } = await import("@tauri-apps/api/path");
+    return join(await appDataDir(), KEY_FILE);
+};
+
+const readKeyFile = async () => {
+    const [{ exists, readFile }, path] = await Promise.all([
+        import("@tauri-apps/plugin-fs"),
+        keyFilePath(),
+    ]);
+    if (!(await exists(path))) return undefined;
+    try {
+        const value = JSON.parse(
+            new TextDecoder().decode(await readFile(path)),
+        ) as { localChatKey?: unknown };
+        return typeof value.localChatKey === "string"
+            ? value.localChatKey
+            : undefined;
+    } catch {
+        return undefined;
+    }
+};
+
+const isLocalChatMigrationDone = () => localGet(MIGRATION_DONE) === "1";
+
+export const promoteLocalChatKey = async (
+    current: string | undefined,
+    currentSecureKey: string,
+    browserKey: string,
+) => {
+    if (isLocalChatMigrationDone()) {
+        _migrationSourceKeys = current ? [current] : [];
+        return current;
+    }
+    const [oldSecure, file] = await Promise.all([
+        secureStorageGet(OLD_LOCAL_SECURE_KEY),
+        readKeyFile(),
+    ]);
+    _migrationSourceKeys = [
+        current,
+        localGet(browserKey),
+        oldSecure,
+        file,
+    ].filter(
+        (key, index, keys): key is string =>
+            !!key && keys.indexOf(key) === index,
+    );
+    const key = _migrationSourceKeys[0];
+    if (key && key !== current) await secureStorageSet(currentSecureKey, key);
+    return key;
+};
+
+export const hasRetiredLocalChatStore = () => !!localGet(LOCAL_STORAGE_STORE);
+
+const retiredUserID = () => {
+    try {
+        const user = JSON.parse(localStorage.getItem("user") ?? "null") as {
+            id?: unknown;
+        } | null;
+        return typeof user?.id === "number" ? user.id : undefined;
+    } catch {
+        return undefined;
+    }
+};
+
+const finalizeLocalChatMigration = async (deleteOldLocalKeys: boolean) => {
+    const userID = retiredUserID();
+    localStorage.removeItem(LOCAL_STORAGE_STORE);
+    const secureKeys = [
+        "masterKey",
+        "remoteChatKey",
+        ...(userID
+            ? [`remoteChatKey.${userID}`, `remoteChatKey.v2.${userID}`]
+            : []),
+    ];
+    if (deleteOldLocalKeys) {
+        secureKeys.push(OLD_LOCAL_SECURE_KEY, LEGACY_ATTACHMENT_SECURE_KEY);
+    }
+    await Promise.all(secureKeys.map(secureStorageDelete));
+
+    if (deleteOldLocalKeys) {
+        const [{ exists, remove }, { appDataDir, join }] = await Promise.all([
+            import("@tauri-apps/plugin-fs"),
+            import("@tauri-apps/api/path"),
+        ]);
+        const keyFile = await join(await appDataDir(), KEY_FILE);
+        if (await exists(keyFile)) await remove(keyFile);
+    }
+
+    for (const key of Object.keys(localStorage)) {
+        if (
+            key.startsWith("ensu.chatKey") &&
+            (deleteOldLocalKeys || key !== OLD_LOCAL_BROWSER_KEY)
+        ) {
+            localStorage.removeItem(key);
+        }
+    }
+    [
+        "user",
+        "srpAttributes",
+        "keyAttributes",
+        "originalKeyAttributes",
+        "isFirstLogin",
+        "ensu.chat.nativeMigration.v2",
+    ].forEach((key) => localStorage.removeItem(key));
+    await removeKV("token");
+    localStorage.setItem(MIGRATION_DONE, "1");
+};
 
 const liveStore = (
     sessions: OldSession[],
@@ -356,12 +471,12 @@ const reencryptAttachments = async (chatKey: string, sourceKeys: string[]) => {
     return allCurrent;
 };
 
-const migrate = async (chatKey: string) => {
+export const openChatStoreWithCompatibility = async (chatKey: string) => {
     if (isTauriRuntime()) {
         const migrationPending = !isLocalChatMigrationDone();
         const sourceKeys = distinctKeys([
             chatKey,
-            ...(migrationPending ? localChatMigrationSourceKeys() : []),
+            ...(migrationPending ? _migrationSourceKeys : []),
         ]);
         await invokeChat("chat_db_open", {
             input: { keyB64: chatKey, recoveryKeysB64: sourceKeys },
@@ -378,17 +493,4 @@ const migrate = async (chatKey: string) => {
         await removeCurrentBrowserTombstones();
         await importOldBrowserStores();
     }
-};
-
-export const initializeChatStorePersistence = async (chatKey: string) => {
-    if (_initPromise) return _initPromise;
-    const promise = migrate(chatKey).catch((error: unknown) => {
-        if (_initPromise === promise) {
-            _initPromise = undefined;
-        }
-        log.error("Failed to initialize chat persistence", error);
-        throw error;
-    });
-    _initPromise = promise;
-    return promise;
 };
